// cls/src/app/enums/e-menus.enum.ts
//所有menu key值
export enum EMenuKeys {
  //待办
  SCHEDULE = "1-1",
  SCHEDULE_LIST = "1-1-1",

  //结算
  SETTLE = "1-23",
  SETTLE_LIST = "1-23-1",

  //订单
  ORDER = "1-24",
  ORDER_LIST = "1-24-1",
  ORDER_ITEM = "1-24-2",

}


//所有menu的名称
export enum EMenuNames {
  SCHEDULE = "首页",
  SCHEDULE_LIST = "待办事项",

  SETTLE = "结算管理",
  SETTLE_LIST = "结算列表",

  ORDER = "订单管理",
  ORDER_LIST = "订单列表",
  ORDER_ITEM = "打印订单"
}